requirements.txt
appdirs==1.4.3
Django==1.10.7
django-braces==1.11.0
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
six==1.10.0
